Revert inconsistent branch fallback
Was introduced in 4a47489 but only for the SAC wagon. By now we have
way better fallback mechanisms in the reusable workflows in the core,
and all the other wagons have been working fine without this fallback.
This way, we get a little bit closer to a consistent setup for all
wagons.
Fixes hitobito/hitobito#2814